pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "signalforge"
version = "0.1.0"
description = "Multi-asset buy/sell signal system leveraging Kronos, Qlib, TradingAgents, and more"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    # Data ingestion
    "yfinance>=0.2.63",
    "ccxt>=4.0",
    "pandas>=2.0",
    "polars>=1.0",
    "pyarrow>=14.0",

    # Technical analysis
    "pandas-ta>=0.3.14b",

    # ML / Deep Learning
    "torch>=2.0",
    "transformers>=4.40",
    "safetensors",
    "scikit-learn>=1.3",

    # Kronos dependencies
    "einops",

    # Configuration & CLI
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",

    # Utilities
    "httpx>=0.25",
    "tqdm",
    "loguru",
]

[project.optional-dependencies]
qlib = ["pyqlib>=0.9.6"]
agents = [
    "tradingagents>=0.2.0",
    "langchain-core>=0.3",
    "langchain-openai",
    "langchain-anthropic",
]
chronos = ["chronos-forecasting>=1.0"]
backtest = ["vectorbt>=0.26"]
sentiment = ["transformers>=4.40"]
all = ["signalforge[qlib,agents,chronos,backtest,sentiment]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
signalforge = "signalforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/signalforge"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
